+++ resources/assets/sass/components/banner.scss
@@ -142,9 +142,6 @@
 .codeweek-banner.about{
   background-color: #72A8D0;
 }
-.codeweek-banner.search{
-  background-color: #164194;
-}
 .codeweek-banner.error{
   background-color: #e57373;
 }
