+++ routes/web.php
@@ -92,7 +92,6 @@ Route::get('/beambassador', [StaticPageController::class, 'static'])->name(
 Route::get('/about', [StaticPageController::class, 'static'])->name('about');
 Route::get('/our-values', [StaticPageController::class, 'static'])->name('our-values');
 Route::get('/partners', [StaticPageController::class, 'static'])->name('sponsors');
-
 Route::get('/codeweek4all', [StaticPageController::class, 'static'])->name(
     'codeweek4all'
 );
@@ -330,9 +329,8 @@ Route::get('/my', [EventController::class, 'my'])
     ->middleware('auth')
     ->name('my_events');
 
-// Route::get('/search', [SearchController::class, 'search'])->name('search_event');
-// Route::post('/search', [SearchController::class, 'searchPOST'])->name('search_events');
-Route::get('/search', [SearchController::class, 'index'])->name('search_event');
+Route::get('/search', [SearchController::class, 'search'])->name('search_event');
+Route::post('/search', [SearchController::class, 'searchPOST'])->name('search_events');
 Route::get('/scoreboard', [ScoreboardController::class, 'index'])->name('scoreboard');
 Route::patch('user', [UserController::class, 'update'])
     ->name('user.update')
